apps/worker/pyproject.toml
[project]
name = "sidekick-worker"
version = "0.1.0"
description = "Sidekick Celery async worker (M2)"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.4",
    "redis>=5.2",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "rich>=13.9",
    "sidekick-agent",
    "sidekick-tools",
    "sidekick-api",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker"]
